scripts/competitor-scraper/requirements.txt
scrapling[all]>=0.2
psycopg2-binary>=2.9
python-dotenv>=1.0
